% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defcountry}
\alias{defcountry}
\title{defcountry}
\format{An object of class \code{data.frame} with 250 rows and 2 columns.}
\usage{
defcountry
}
\description{
Country list
}
\keyword{datasets}
