% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defspace}
\alias{defspace}
\title{defspace}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 8396 rows and 5 columns.}
\usage{
defspace
}
\description{
Spatial definition of spatial object
}
\keyword{datasets}
