% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classXXXMethods.R
\docType{methods}
\name{dim,Time-method}
\alias{dim,Time-method}
\title{generic dim by class}
\usage{
\S4method{dim}{Time}(x)

\S4method{dim}{Time}(x)
}
\description{
generic dim by class

generic dim by class
}
