% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classXXXMethods.R
\docType{methods}
\name{dim,Vessel-method}
\alias{dim,Vessel-method}
\title{generic dim by class}
\usage{
\S4method{dim}{Vessel}(x)
}
\arguments{
\item{x}{an Object}
}
\description{
generic dim by class
}
