% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classXXXMethods.R
\name{dimclass}
\alias{dimclass}
\title{dimension}
\usage{
dimclass(Object)
}
\arguments{
\item{Object}{an object}

\item{...}{the object or parameters from which inheritance is transmitted}
}
\value{
the update object
}
\description{
a generic function to have the dimensiob object
}
