% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classXXXMethods.R
\name{dimslots}
\alias{dimslots}
\title{dimension of slots}
\usage{
dimslots(Object)
}
\arguments{
\item{Object}{an object}

\item{...}{the object or parameters from which inheritance is transmitted}
}
\value{
the update object
}
\description{
a generic function to have the length of each slot in an object
}
