% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00function.R
\name{importSlots}
\alias{importSlots}
\title{Import slot of an object to another object}
\usage{
importSlots(object1, object2)
}
\arguments{
\item{object1}{an object to be transfered to object2}

\item{object2}{the object in which object1 will be copied}
}
\value{
the updated object2
}
\description{
Import slot of an object to another object
}
