% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class0Sampling.R
\docType{class}
\name{Sampling-class}
\alias{Sampling-class}
\title{Class Sampling}
\description{
Class Sampling
}
\section{Slots}{

\describe{
\item{\code{type}}{}

\item{\code{SamplingMethod}}{}

\item{\code{SamplingProject}}{}

\item{\code{SamplingCountry}}{}
}}

