% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00data.R
\docType{data}
\name{Samplingmethodtype}
\alias{Samplingmethodtype}
\title{Samplingmethodtype}
\format{An object of class \code{character} of length 2.}
\usage{
Samplingmethodtype
}
\description{
Sampling method type
}
\keyword{datasets}
