% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class1Trip.R
\docType{class}
\name{Trip-class}
\alias{Trip-class}
\title{Class Trip}
\description{
Class Trip
}
\section{Slots}{

\describe{
\item{\code{vessel}}{}

\item{\code{sampling}}{}

\item{\code{time}}{}

\item{\code{space}}{}

\item{\code{nbhaul}}{}

\item{\code{daysatsea}}{}
}}

