% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00data.R
\docType{data}
\name{Tripmethodtype}
\alias{Tripmethodtype}
\title{Tripmethodtype}
\format{An object of class \code{character} of length 4.}
\usage{
Tripmethodtype
}
\description{
Trip method type
}
\keyword{datasets}
