% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class0Sampling.R
\name{validSampling}
\alias{validSampling}
\title{validity Sampling SamplingMethod}
\usage{
validSampling(object)
}
\arguments{
\item{object}{a time object}
}
\description{
validity Sampling SamplingMethod
}
