% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class0Vessel.R
\docType{class}
\name{Vessel-class}
\alias{Vessel-class}
\title{Class Vessel}
\description{
Class Vessel
}
\section{Slots}{

\describe{
\item{\code{VesselId}}{}

\item{\code{VesselFlag}}{}

\item{\code{VesselType}}{}

\item{\code{VesselLength}}{}

\item{\code{VesselPower}}{}

\item{\code{VesselSize}}{}
}}

